Make the build reproducible
help2man was vendored as part of d5d5253 (Fix bugs reported by
Dominique, 2000-10-06) because at the time, help2man did not have a
--version-option argument and one needed to be patched in. However, as
of help2man 1.29, released in July 2003, this option was added.

Remove the vendored version of help2man and use the help2man on the
system. It's been released for long enough that all users should be
using a version of help2man that's less than 22 years old.

This helps because distros such as Arch Linux patch the vendored version
of help2man in order to make the build reproducible[0]. However, help2man
has been updated to support $SOURCE_DATE_EPOCH for a long time now.

Add -n to the gzip call so that the timestamp of the files are not
saved.

While we're at it, add the generated files to the .gitignore.

[0]: https://gitlab.archlinux.org/archlinux/packaging/packages/lsb-release/-/blob/e5ce86591a488e1089e1c246a296adc70679588b/lsb_release_make_man_page_reproducible.patch